fix(tpcc): apply PACING to procs variant; dedupe tx/procs via tpcc_common
PACING=true had no effect on tpcc/procs: the keying+think-time pacing code
lived only in tx.ts, so stored-procedure runs ignored it and ran unpaced.

Move the shared scaffolding (driver setup, SQL parse, isolation, retry policy,
prepare lifecycle, §5.2.5 pacing, weighted dispatch, post-run summary) into
tpcc_common.ts, mirroring the tpcb layout. tx.ts and procs.ts now hold only
their transaction bodies and import the rest. Net -213 lines across the three
files.

Both variants now run with errorMode=throw (previously procs threw while tx
logged); tx already threw on failed queries via its if-no-row guards, so the
observable behaviour is unchanged apart from the propagated error message.

Closes #113
